requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.10
python-dotenv==1.0.0
pydantic==2.4.2
urllib3==2.3.0
bs4==0.0.2
certifi==2025.1.31 